refactor(daemon): co-locate auth guards in route files
Move all permission/rate-limit middleware from daemon.ts centralized
guard block into their respective route registration functions. Each
route file now owns its auth guards, eliminating the ~207-line guard
block in daemon.ts and fixing double-execution of guards in
memory-routes.ts.

13 files changed: 12 route files gained guards, daemon.ts lost the
centralized block and unused imports.
